requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sarvamai>=0.1.11
motor>=3.3.0
pymongo>=4.6.0
httpx>=0.27.0
python-multipart
websockets
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
redis>=5.0.0
yt-dlp>=2025.1.15
json-repair>=0.45.0

deepgram-sdk==3.*

aiohttp
